mds: fix trim with count
There was an off-by-one issue here because of checking
lru_get_size after calling lru_expire (which decrements
the lru's size).

It manifested during exporting a dir because we would
trim one dentry too few at the end of an export.

We should also push dn back onto the lru if we're dropping
out at this stage.

Signed-off-by: John Spray <john.spray@redhat.com>